Following most of Thomas proposals
